# src/app.js
import { readFileSync } from 'node:fs'
import { dirname, join } from 'node:path'
import { fileURLToPath } from 'node:url'
import Fastify from 'fastify'
import cors from '@fastify/cors'
import rateLimit from '@fastify/rate-limit'
import { randomUUID, timingSafeEqual } from 'node:crypto'
import { appendQuote, findQuoteById, listQuotesRecent, checkStorageReady, countQuotes, updateQuoteById, quoteStatusBreakdown } from './store.js'
import { appendLead, listLeadsRecent, countLeads, findLeadById, updateLeadById, leadStatusBreakdown } from './leadsStore.js'

const PROJECT_TYPES = ['landing', 'website', 'dashboard']
const ADDON_IDS = ['design', 'i18n', 'rush']
const QUOTE_STATUSES = ['draft', 'sent', 'accepted', 'declined']
const LEAD_STATUSES = ['new', 'contacted', 'qualified', 'closed']
const LEAD_SOURCES = ['landing', 'calculator']

/** RFC 9562 UUID v1–v5 shape (ids from `randomUUID`). */
const UUID_PARAM =
  /^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i

/** JSON Schema for `POST /api/v1/quotes` (AJV). */
const postQuoteBodySchema = {
  type: 'object',
  required: ['projectType', 'addOnIds', 'min', 'max'],
  properties: {
    projectType: { type: 'string', enum: PROJECT_TYPES },
    addOnIds: {
      type: 'array',
      items: { type: 'string', enum: ADDON_IDS },
    },
    extraSections: { anyOf: [{ type: 'string' }, { type: 'number' }] },
    min: { anyOf: [{ type: 'number' }, { type: 'string' }] },
    max: { anyOf: [{ type: 'number' }, { type: 'string' }] },
    lang: { enum: ['en', 'zh'] },
    quoteRef: { anyOf: [{ type: 'string' }, { type: 'null' }] },
    summary: { anyOf: [{ type: 'string' }, { type: 'null' }] },
  },
  additionalProperties: false,
}

const postLeadBodySchema = {
  type: 'object',
  required: ['name', 'email', 'message'],
  properties: {
    name: { type: 'string', minLength: 1, maxLength: 200 },
    email: { type: 'string', minLength: 3, maxLength: 320 },
    message: { type: 'string', minLength: 1, maxLength: 8000 },
    subject: { anyOf: [{ type: 'string' }, { type: 'null' }] },
    projectType: { anyOf: [{ type: 'string' }, { type: 'null' }] },
    timeline: { anyOf: [{ type: 'string' }, { type: 'null' }] },
    quoteRef: { anyOf: [{ type: 'string' }, { type: 'null' }] },
    source: { enum: LEAD_SOURCES },
    lang: { enum: ['en', 'zh'] },
  },
  additionalProperties: false,
}

const patchQuoteBodySchema = {
  type: 'object',
  required: ['status'],
  properties: {
    status: { type: 'string', enum: QUOTE_STATUSES },
  },
  additionalProperties: false,
}

const patchLeadBodySchema = {
  type: 'object',
  required: ['status'],
  properties: {
    status: { type: 'string', enum: LEAD_STATUSES },
  },
  additionalProperties: false,
}

const __dirname = dirname(fileURLToPath(import.meta.url))
const pkg = JSON.parse(
  readFileSync(join(__dirname, '..', 'package.json'), 'utf8')
)

function safeTokenEquals(a, b) {
  const aa = Buffer.from(a)
  const bb = Buffer.from(b)
  return aa.length === bb.length && timingSafeEqual(aa, bb)
}

function buildQuoteShareLinks(id) {
  const calc = (
    process.env.FRONTEND_CALCULATOR_URL ||
    'https://pixellayer7-jpg.github.io/project-estimator/'
  ).replace(/\/?$/, '/')
  const landing = (
    process.env.FRONTEND_LANDING_URL || 'https://pixellayer7-jpg.github.io/1/'
  ).replace(/\/?$/, '/')
  return {
    calculator: `${calc}?load=${encodeURIComponent(id)}`,
    contact: `${landing}?quote=${encodeURIComponent(id)}#contact`,
  }
}

function getBearerToken(request) {
  const header = request.headers.authorization
  if (typeof header !== 'string') return ''
  const match = header.match(/^Bearer\s+(.+)$/i)
  return match ? match[1].trim() : ''
}

function authorizeQuoteList(request, reply) {
  const expected = String(process.env.LIST_QUOTES_TOKEN || '').trim()
  if (!expected) return true
  const provided = getBearerToken(request)
  if (provided && safeTokenEquals(provided, expected)) return true
  reply
    .header('WWW-Authenticate', 'Bearer')
    .code(401)
    .send({ error: 'Unauthorized' })
  return false
}

export function buildOpenApiDocument(version) {
  const bearer = [{ bearerAuth: [] }]
  const uuidParam = {
    name: 'id',
    in: 'path',
    required: true,
    schema: { type: 'string', format: 'uuid' },
    description: 'RFC 9562 UUID',
  }
  const quoteExample = {
    projectType: 'landing',
    addOnIds: ['i18n'],
    extraSections: '2',
    min: 800,
    max: 1400,
    lang: 'en',
    summary: 'Interview demo quote',
  }
  const leadExample = {
    name: 'Ada Interview',
    email: 'ada@example.com',
    message: 'Need a bilingual landing page',
    source: 'landing',
    lang: 'en',
  }
  const quoteCreateResponse = {
    id: 'aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee',
    createdAt: '2026-09-16T12:00:00.000Z',
    path: '/api/v1/quotes/aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee',
    loadQuery: '?load=aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee',
    links: {
      calculator:
        'https://pixellayer7-jpg.github.io/project-estimator/?load=aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee',
      contact:
        'https://pixellayer7-jpg.github.io/1/?quote=aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee#contact',
    },
  }
  const statsExample = {
    totalQuotes: 3,
    totalLeads: 2,
    quotesByStatus: { draft: 1, sent: 1, accepted: 1, declined: 0 },
    leadsByStatus: { new: 1, contacted: 1, qualified: 0, closed: 0 },
    version,
  }

  return {
    openapi: '3.0.3',
    info: {
      title: 'PixelLayer estimator-api',
      version,
      description:
        'File-backed quote + lead CRM for PixelLayer. Public GET by quote UUID for share links; Bearer protects list/PATCH when LIST_QUOTES_TOKEN is set.',
      contact: {
        name: 'PixelLayer L.L.C',
        email: 'pixellayer7@gmail.com',
        url: 'https://pixellayer7-jpg.github.io/1/',
      },
    },
    servers: [
      { url: 'http://localhost:3000', description: 'Local dev' },
      {
        url: 'https://your-estimator-api.example.com',
        description: 'Deployed API (Render / Railway / Fly)',
      },
    ],
    tags: [
      { name: 'System' },
      { name: 'Quotes' },
      { name: 'Leads' },
    ],
    components: {
      securitySchemes: {
        bearerAuth: {
          type: 'http',
          scheme: 'bearer',
          bearerFormat: 'opaque',
          description:
            'Same value as LIST_QUOTES_TOKEN. Required for list/PATCH (and GET lead) when the env var is set.',
        },
      },
      schemas: {
        QuoteCreate: postQuoteBodySchema,
        QuoteStatusPatch: patchQuoteBodySchema,
        LeadCreate: postLeadBodySchema,
        LeadStatusPatch: patchLeadBodySchema,
        Error: {
          type: 'object',
          properties: { error: { type: 'string' } },
          required: ['error'],
        },
        Stats: {
          type: 'object',
          properties: {
            totalQuotes: { type: 'number' },
            totalLeads: { type: 'number' },
            quotesByStatus: { type: 'object', additionalProperties: { type: 'number' } },
            leadsByStatus: { type: 'object', additionalProperties: { type: 'number' } },
            version: { type: 'string' },
          },
        },
      },
      examples: {
        QuoteCreateExample: { summary: 'Landing + i18n', value: quoteExample },
        LeadCreateExample: { summary: 'Landing contact', value: leadExample },
        QuoteCreated: { summary: '201 response', value: quoteCreateResponse },
        StatsExample: { summary: 'Counts + breakdown', value: statsExample },
      },
    },
    paths: {
      '/': {
        get: {
          tags: ['System'],
          summary: 'Service catalog',
          responses: {
            200: {
              description: 'Name, version, endpoint map',
            },
          },
        },
      },
      '/health': {
        get: {
          tags: ['System'],
          summary: 'Health + storage check',
          responses: {
            200: { description: 'ok + storage ready' },
            503: { description: 'Storage unavailable' },
          },
        },
      },
      '/api/v1/openapi.json': {
        get: {
          tags: ['System'],
          summary: 'This OpenAPI document',
          responses: { 200: { description: 'OpenAPI 3.0.3 JSON' } },
        },
      },
      '/api/v1/stats': {
        get: {
          tags: ['System'],
          summary: 'Quote and lead counts with status breakdown',
          responses: {
            200: {
              description: 'Aggregates',
              content: {
                'application/json': {
                  schema: { $ref: '#/components/schemas/Stats' },
                  examples: {
                    sample: { $ref: '#/components/examples/StatsExample' },
                  },
                },
              },
            },
          },
        },
      },
      '/api/v1/quotes': {
        get: {
          tags: ['Quotes'],
          summary: 'List recent quotes (optional Bearer)',
          security: bearer,
          parameters: [
            {
              name: 'limit',
              in: 'query',
              schema: { type: 'integer', minimum: 1, maximum: 100, default: 20 },
            },
          ],
          responses: {
            200: { description: '{ count, items } — items omit summary' },
            401: {
              description: 'Unauthorized when LIST_QUOTES_TOKEN is set',
              content: {
                'application/json': {
                  schema: { $ref: '#/components/schemas/Error' },
                },
              },
            },
          },
        },
        post: {
          tags: ['Quotes'],
          summary: 'Create quote snapshot',
          requestBody: {
            required: true,
            content: {
              'application/json': {
                schema: { $ref: '#/components/schemas/QuoteCreate' },
                examples: {
                  landing: { $ref: '#/components/examples/QuoteCreateExample' },
                },
              },
            },
          },
          responses: {
            201: {
              description: 'Created — includes loadQuery and share links',
              content: {
                'application/json': {
                  examples: {
                    created: { $ref: '#/components/examples/QuoteCreated' },
                  },
                },
              },
            },
            400: {
              description: 'Invalid body / min>max / bad quoteRef',
              content: {
                'application/json': {
                  schema: { $ref: '#/components/schemas/Error' },
                },
              },
            },
            413: { description: 'Body larger than 256 KiB' },
          },
        },
      },
      '/api/v1/quotes/{id}': {
        get: {
          tags: ['Quotes'],
          summary: 'Get quote by UUID (public share link)',
          parameters: [uuidParam],
          responses: {
            200: { description: 'Quote record' },
            400: {
              description: 'Invalid id',
              content: {
                'application/json': {
                  schema: { $ref: '#/components/schemas/Error' },
                },
              },
            },
            404: {
              description: 'Not found',
              content: {
                'application/json': {
                  schema: { $ref: '#/components/schemas/Error' },
                },
              },
            },
          },
        },
        patch: {
          tags: ['Quotes'],
          summary: 'Update quote status',
          security: bearer,
          parameters: [uuidParam],
          requestBody: {
            required: true,
            content: {
              'application/json': {
                schema: { $ref: '#/components/schemas/QuoteStatusPatch' },
                examples: {
                  sent: { value: { status: 'sent' } },
                },
              },
            },
          },
          responses: {
            200: { description: 'Updated quote' },
            400: { description: 'Invalid id or body' },
            401: { description: 'Unauthorized when token set' },
            404: { description: 'Not found' },
          },
        },
      },
      '/api/v1/leads': {
        get: {
          tags: ['Leads'],
          summary: 'List recent leads (optional Bearer)',
          security: bearer,
          parameters: [
            {
              name: 'limit',
              in: 'query',
              schema: { type: 'integer', minimum: 1, maximum: 100, default: 20 },
            },
          ],
          responses: {
            200: { description: '{ count, items }' },
            401: { description: 'Unauthorized when token set' },
          },
        },
        post: {
          tags: ['Leads'],
          summary: 'Create contact lead',
          requestBody: {
            required: true,
            content: {
              'application/json': {
                schema: { $ref: '#/components/schemas/LeadCreate' },
                examples: {
                  landing: { $ref: '#/components/examples/LeadCreateExample' },
                },
              },
            },
          },
          responses: {
            201: { description: 'Created — { id, createdAt, path }' },
            400: { description: 'Invalid body / quoteRef' },
          },
        },
      },
      '/api/v1/leads/{id}': {
        get: {
          tags: ['Leads'],
          summary: 'Get lead by UUID',
          security: bearer,
          parameters: [uuidParam],
          responses: {
            200: { description: 'Lead record' },
            400: { description: 'Invalid id' },
            401: { description: 'Unauthorized when token set' },
            404: { description: 'Not found' },
          },
        },
        patch: {
          tags: ['Leads'],
          summary: 'Update lead status',
          security: bearer,
          parameters: [uuidParam],
          requestBody: {
            required: true,
            content: {
              'application/json': {
                schema: { $ref: '#/components/schemas/LeadStatusPatch' },
                examples: {
                  contacted: { value: { status: 'contacted' } },
                },
              },
            },
          },
          responses: {
            200: { description: 'Updated lead' },
            400: { description: 'Invalid id or body' },
            401: { description: 'Unauthorized when token set' },
            404: { description: 'Not found' },
          },
        },
      },
    },
  }
}

export default async function buildApp() {
  const app = Fastify({ logger: false, bodyLimit: 262_144 })

  const origins = (process.env.CORS_ORIGIN || '')
    .split(',')
    .map((s) => s.trim())
    .filter(Boolean)

  await app.register(cors, {
    origin: origins.length ? origins : true,
  })

  await app.register(rateLimit, {
    max: 100,
    timeWindow: '1 minute',
  })

  app.addHook('onSend', async (_request, reply, payload) => {
    reply.header('X-Content-Type-Options', 'nosniff')
    reply.header('X-API-Version', pkg.version)
    return payload
  })

  app.setErrorHandler((error, _request, reply) => {
    if (error.validation) {
      return reply.code(400).send({ error: 'Invalid request body' })
    }
    if (error.statusCode) {
      return reply.code(error.statusCode).send({
        error: typeof error.message === 'string' ? error.message : 'Error',
      })
    }
    return reply.code(500).send({ error: 'Internal Server Error' })
  })

  app.get('/', async () => ({
    service: 'estimator-api',
    version: pkg.version,
    links: {
      landing: 'https://pixellayer7-jpg.github.io/1/',
      calculator: 'https://pixellayer7-jpg.github.io/project-estimator/',
    },
    endpoints: {
      health: '/health',
      listQuotes: 'GET /api/v1/quotes?limit=20',
      createQuote: 'POST /api/v1/quotes',
      getQuote: 'GET /api/v1/quotes/:id',
      patchQuote: 'PATCH /api/v1/quotes/:id',
      createLead: 'POST /api/v1/leads',
      listLeads: 'GET /api/v1/leads?limit=20',
      getLead: 'GET /api/v1/leads/:id',
      patchLead: 'PATCH /api/v1/leads/:id',
      openapi: 'GET /api/v1/openapi.json',
      stats: 'GET /api/v1/stats',
    },
  }))

  app.get('/api/v1/openapi.json', async () => buildOpenApiDocument(pkg.version))

  app.get('/api/v1/stats', async () => {
    const [totalQuotes, totalLeads, quotesByStatus, leadsByStatus] =
      await Promise.all([
        countQuotes(),
        countLeads(),
        quoteStatusBreakdown(),
        leadStatusBreakdown(),
      ])
    return {
      totalQuotes,
      totalLeads,
      quotesByStatus,
      leadsByStatus,
      version: pkg.version,
    }
  })

  app.get('/health', async (_request, reply) => {
    try {
      await checkStorageReady()
      return { ok: true, service: 'estimator-api', version: pkg.version, storage: 'ready' }
    } catch (e) {
      return reply.code(503).send({
        ok: false,
        service: 'estimator-api',
        version: pkg.version,
        storage: 'unavailable',
        error: e instanceof Error ? e.message : 'Storage check failed',
      })
    }
  })

  app.get('/api/v1/quotes', async (request, reply) => {
    if (!authorizeQuoteList(request, reply)) return
    let limit = 20
    const raw = request.query.limit
    if (raw !== undefined && raw !== '') {
      const p = Number.parseInt(String(raw), 10)
      if (Number.isFinite(p)) limit = p
    }
    const items = await listQuotesRecent(limit)
    return { count: items.length, items }
  })

  app.post(
    '/api/v1/quotes',
    {
      config: { rateLimit: { max: 30, timeWindow: '1 minute' } },
      schema: { body: postQuoteBodySchema },
    },
    async (request, reply) => {
      const body = request.body
      const {
        projectType,
        addOnIds,
        extraSections,
        min,
        max,
        lang,
        quoteRef,
        summary,
      } = body

      const minNum = typeof min === 'number' ? min : Number(min)
      const maxNum = typeof max === 'number' ? max : Number(max)
      if (!Number.isFinite(minNum) || !Number.isFinite(maxNum)) {
        return reply
          .code(400)
          .send({ error: 'min and max must be finite numbers' })
      }
      if (minNum > maxNum) {
        return reply
          .code(400)
          .send({ error: 'min must be less than or equal to max' })
      }
      if (
        typeof quoteRef === 'string' &&
        quoteRef.length > 0 &&
        !UUID_PARAM.test(quoteRef)
      ) {
        return reply.code(400).send({ error: 'Invalid quoteRef' })
      }

      const id = randomUUID()
      const createdAt = new Date().toISOString()
      const extraStoredRaw =
        extraSections === undefined || extraSections === null
          ? 0
          : Number(extraSections)
      const extraClamped = Number.isFinite(extraStoredRaw)
        ? Math.min(20, Math.max(0, Math.floor(extraStoredRaw)))
        : 0
      const extraStored = String(extraClamped)

      const record = {
        id,
        createdAt,
        projectType,
        addOnIds,
        extraSections: extraStored,
        min: minNum,
        max: maxNum,
        lang: typeof lang === 'string' ? lang : 'en',
        quoteRef: typeof quoteRef === 'string' ? quoteRef : null,
        summary: typeof summary === 'string' ? summary : null,
        status: 'draft',
      }

      await appendQuote(record)
      return reply.code(201).send({
        id,
        createdAt,
        path: `/api/v1/quotes/${id}`,
        loadQuery: `?load=${id}`,
        links: buildQuoteShareLinks(id),
      })
    }
  )

  app.get('/api/v1/quotes/:id', async (request, reply) => {
    const { id } = request.params
    if (!UUID_PARAM.test(id)) {
      return reply.code(400).send({ error: 'Invalid id' })
    }
    const row = await findQuoteById(id)
    if (!row) return reply.code(404).send({ error: 'Not found' })
    return row
  })

  app.patch(
    '/api/v1/quotes/:id',
    { schema: { body: patchQuoteBodySchema } },
    async (request, reply) => {
      if (!authorizeQuoteList(request, reply)) return
      const { id } = request.params
      if (!UUID_PARAM.test(id)) {
        return reply.code(400).send({ error: 'Invalid id' })
      }
      const row = await findQuoteById(id)
      if (!row) return reply.code(404).send({ error: 'Not found' })
      const updated = await updateQuoteById(id, { status: request.body.status })
      return updated
    }
  )

  app.get('/api/v1/leads', async (request, reply) => {
    if (!authorizeQuoteList(request, reply)) return
    let limit = 20
    const raw = request.query.limit
    if (raw !== undefined && raw !== '') {
      const p = Number.parseInt(String(raw), 10)
      if (Number.isFinite(p)) limit = p
    }
    const items = await listLeadsRecent(limit)
    return { count: items.length, items }
  })

  app.post(
    '/api/v1/leads',
    {
      config: { rateLimit: { max: 20, timeWindow: '1 minute' } },
      schema: { body: postLeadBodySchema },
    },
    async (request, reply) => {
      const body = request.body
      const { name, email, message, subject, projectType, timeline, quoteRef, source, lang } =
        body

      if (
        typeof quoteRef === 'string' &&
        quoteRef.length > 0 &&
        !UUID_PARAM.test(quoteRef)
      ) {
        return reply.code(400).send({ error: 'Invalid quoteRef' })
      }

      const id = randomUUID()
      const createdAt = new Date().toISOString()
      const record = {
        id,
        createdAt,
        name: String(name).trim(),
        email: String(email).trim(),
        message: String(message).trim(),
        subject: typeof subject === 'string' ? subject.trim() : null,
        projectType: typeof projectType === 'string' ? projectType.trim() : null,
        timeline: typeof timeline === 'string' ? timeline.trim() : null,
        quoteRef: typeof quoteRef === 'string' ? quoteRef.trim() : null,
        source: typeof source === 'string' ? source : 'landing',
        lang: typeof lang === 'string' ? lang : 'en',
        status: 'new',
      }

      await appendLead(record)
      return reply.code(201).send({ id, createdAt, path: `/api/v1/leads/${id}` })
    }
  )

  app.get('/api/v1/leads/:id', async (request, reply) => {
    if (!authorizeQuoteList(request, reply)) return
    const { id } = request.params
    if (!UUID_PARAM.test(id)) {
      return reply.code(400).send({ error: 'Invalid id' })
    }
    const row = await findLeadById(id)
    if (!row) return reply.code(404).send({ error: 'Not found' })
    return row
  })

  app.patch(
    '/api/v1/leads/:id',
    { schema: { body: patchLeadBodySchema } },
    async (request, reply) => {
      if (!authorizeQuoteList(request, reply)) return
      const { id } = request.params
      if (!UUID_PARAM.test(id)) {
        return reply.code(400).send({ error: 'Invalid id' })
      }
      const row = await findLeadById(id)
      if (!row) return reply.code(404).send({ error: 'Not found' })
      const updated = await updateLeadById(id, { status: request.body.status })
      return updated
    }
  )

  app.setNotFoundHandler((_request, reply) => {
    reply.code(404).send({ error: 'Not found' })
  })

  return app
}